fixed context loss issue
state from context kept getting lose in hard refresh of the results and character pages. to get around this, I created a failsafe by running a request to the api with the value in the slug using the context object from next.js. So the first data that will be used is from what is stored in context. if that is not available, it will query the server for the data